Rework caching of `ImmutableAttributes`
This commit completly reworks the storage of attributes and immutable attributes. At the root of
the change is the `ImmutableAttributesFactory`, which is used to create, combine and cache
immutable attribute sets. This factory lives at the build scope level, and does *not* use a `Map`
to represent attributes. Instead, it creates a "tree" of attributes, which are mutated by adding
a new attribute. This dramatically reduces the memory pressure caused by the creation of immutable
attributes, as long at the set of attributes and values is small (which should be the case).

It's worth noting that the factory is not a _perfect_ cache. For example, depending on the order
of attributes, several "sets", corresponding to several "path to root" can be created. For example,
we can see `(foo, bar, baz)` and `(baz, foo, bar)`. Even if they are equivalent, there will be
2 distinct paths in the cache. This is not a problem, since equality check works, and the overhead
of de-duplicating is significant, so doesn't bring any advantage.

The commit also introduces a `lockAttributes` method on `ConfigurationInternal`, which allows
effectively locking the attribute set of a configuration, transforming it to an immutable
attribute set. This avoids, if multiple consumers ask for the configuration attributes in an
immutable form, to do the conversion several times (thus reducing the pressure again).

Last but not least, `DefaultAttributesContainer` is the mutable version of the attribute
container, and relies internally on `ImmutableAttributes`. This allows us to keep the semantics
of mutable attributes during the configuration phase (and, in particular, changes the attributes
of a configuration after the artifacts are created should be visible to the artifact attributes),
while deduplicating attributes and keeping memory usage low at the same time.

The drawback of this approach is the general complexity of the code, and having to pass the
`ImmutableAttributesFactory` along.